+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # @project-version@ (@build-time@)
 
 * Improved code for updating nameplates in arenas and battleground, e.g., after a new round in solo shuffles starts.
+* Nameplates for interactable, but dead units now automatically switch to headline view (to hide the empty healthbar) [Comment #7100, GH-485].